+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 1.0.1 (2022.05.31)
+### BUG FIX
+- Missing v in File Version string
+- Type error in load table
+- Typos
+### ENHANCEMENTS
+- Added additional tests
 ## 1.0.0 (2022.05.29)
 
 - Program Released